+++ managers.tf
@@ -23,24 +23,14 @@ data "cloudinit_config" "managers" {
   base64_encode = "true"
 
   part {
-    content = file("${path.module}/common.cloud-config")
-  }
-
-  part {
-    filename = "bootcmd.cloud-config"
-    content = yamlencode({
-      "bootcmd" : [
-        ["cloud-init-per", "once", "amazon-linux-extras-docker", "amazon-linux-extras", "install", "docker"],
-        ["cloud-init-per", "once", "amazon-linux-extras-epel", "amazon-linux-extras", "install", "epel"],
-        ["cloud-init-per", "boot", "/opt/aws/amazon-cloudwatch-agent/bin/amazon-cloudwatch-agent-ctl", "-a", "fetch-config", "-m", "ec2", "-s", "-c", "ssm:${local.cloudwatch_agent_parameter}"],
-      ]
+    content = templatefile("${path.module}/common.cloud-config.tmpl", {
+      cloudwatch_agent_parameter = local.cloudwatch_agent_parameter
     })
-    content_type = "text/cloud-config"
   }
 
   part {
     filename = "ssh_keys.cloud-config"
-    content = var.generate_host_keys ? yamlencode({
+    content = join("\n", ["#cloud-config", var.generate_host_keys ? yamlencode({
       "ssh_keys" : {
         "rsa_private" : tls_private_key.managers-rsa[count.index].private_key_openssh,
         "rsa_public" : tls_private_key.managers-rsa[count.index].public_key_openssh,
@@ -53,7 +43,7 @@ data "cloudinit_config" "managers" {
       "ssh" : {
         "emit_keys_to_console" : false
       }
-    }) : ""
+    }) : ""])
     content_type = "text/cloud-config"
   }
 
@@ -65,15 +55,7 @@ data "cloudinit_config" "managers" {
   }
 
   part {
-    filename     = "init_boto.sh"
-    content      = <<EOT
-#!/bin/sh
-pip3 install botocore boto3
-EOT
-    content_type = "text/x-shellscript"
-  }
-  part {
-    filename = "init_node.py"
+    filename = "init_manager.py"
     content = templatefile(
       "${path.module}/init_node.py",
       {
@@ -118,6 +100,10 @@ resource "aws_instance" "managers" {
   iam_instance_profile = aws_iam_instance_profile.ec2.name
   user_data_base64     = data.cloudinit_config.managers[count.index].rendered
   key_name             = var.key_name
+
+  private_dns_name_options {
+    hostname_type = "resource-name"
+  }
 
   tags = merge({
     Name = "${var.name} manager ${count.index}"